airflow_pipeline/requirements.txt
typer==0.7.0
tensorflow==2.11.0
pandas==1.1.5
wandb==0.15.0
joblib==1.1.1
python-dotenv==0.21.1